requirement.txt
annotated-types==0.7.0
anyio==4.13.0
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
Brotli==1.1.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
colorama==0.4.6
cryptography==46.0.1
cssselect2==0.8.0
distro==1.9.0
Django==5.2.3
django-jalali==7.4.0
djangorestframework==3.16.1
et_xmlfile==2.0.0
fonttools==4.60.0
freetype-py==2.5.1
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jalali_core==1.0.0
jdatetime==5.2.0
jiter==0.13.0
kavenegar==1.1.2
lxml==6.0.2
Markdown==3.9
openai==2.30.0
openpyxl==3.1.5
oscrypto==1.3.0
pillow==11.2.1
psycopg2-binary==2.9.10
pycairo==1.28.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydyf==0.11.0
pyHanko==0.31.0
pyhanko-certvalidator==0.29.0
pypdf==6.1.0
pyphen==0.17.2
python-bidi==0.6.6
pytz==2025.2
PyYAML==6.0.2
reportlab==4.4.4
requests==2.32.4
rlPyCairo==0.4.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
svglib==1.6.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.5.0
weasyprint==66.0
webencodings==0.5.1
xhtml2pdf==0.2.17
zopfli==0.2.3.post1
